// apps/console/src/__generated__/core/MeasureFormDialogCreateMutation.graphql.ts
export type CreateMeasureInput = {
  category: string;
  description?: string | null | undefined;
  name: string;
  organizationId: string;
};
export type MeasureFormDialogCreateMutation$variables = {
  connections: ReadonlyArray<string>;
  input: CreateMeasureInput;
};
export type MeasureFormDialogCreateMutation$data = {
  readonly createMeasure: {
    readonly measureEdge: {
      readonly node: {
        readonly " $fragmentSpreads": FragmentRefs<"MeasureFormDialogMeasureFragment">;
      };
    };
  };
};
export type MeasureFormDialogCreateMutation = {
  response: MeasureFormDialogCreateMutation$data;
  variables: MeasureFormDialogCreateMutation$variables;
};